+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 * Add affect loading progress indicator (`OSIDB-4631`)
 * `Create Jira Task` option not working if there are no changes (`OSIDB-4632`)
+* Offer Aegis-suggested statement text (`AEGIS-XXX`)
 
 ### Changed
 * Removed Affects V1 implementation (`OSIDB-4660`)